+++ actionpack/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Fix `ActionController::Parameters==`
+
+   *Jon Moss*
+
 *   Response etags to always be weak: Prefixes 'W/' to value returned by
    `ActionDispatch::Http::Cache::Response#etag=`, such that etags set in
    `fresh_when` and `stale?` are weak.